# docs/enterprise-digital-twin/contracts/openapi/enterprise-digital-twin.openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise Digital Twin API
  version: 1.0.0
  summary: Permission-aware graph, simulation, governed-action, and synthetic physical-asset API.
  description: Tenant context is always derived from the authenticated principal and a server-issued context handle. Tenant identifiers supplied in payloads are consistency assertions and never select authorization scope. Raw X-Tenant-ID is rejected.
...
servers:
  - url: https://api.enterprise-digital-twin.example
security:
  - oidc: [edt.read]
    contextHeader: []
  - oidc: [edt.read]
    contextCookie: []
...
paths:
...
  /v1/simulations:
    post:
      tags: [Simulation]
      operationId: runSimulation
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [scenario_id]
              properties: {scenario_id: {type: string, format: uuid}}
      responses:
        '202': {description: Durable simulation accepted with recoverable resource and event links., content: {application/json: {schema: {$ref: '#/components/schemas/SimulationRunAccepted'}}}}
        default: {$ref: '#/components/responses/Problem'}
...
  /v1/ai/status:
    get:
      tags: [AI Intelligence]
      operationId: getAiGatewayStatus
      description: Reports provider, durable-store, vector-retrieval, and bounded-agent readiness without exposing credentials or accepting a provider selector.
      x-edt-required-capabilities: [authenticated_membership]
      responses:
        '200': {description: AI gateway readiness; a degraded response is healthy but one or more optional provider capabilities are unavailable., headers: {Cache-Control: {schema: {type: string, const: 'private, no-store'}}}, content: {application/json: {schema: {$ref: '#/components/schemas/AiStatus'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/ai/activity:
    get:
      tags: [AI Intelligence]
      operationId: listAiActivity
      x-edt-authorization-any-of: [connector.admin, 'evidence.read.*']
      parameters: [{$ref: '#/components/parameters/AiPageSize'}]
      responses:
        '200': {description: Actor-scoped redacted AI activity metadata; AI reviewers and administrators may receive the tenant review queue only through separately authorized operations., headers: {Cache-Control: {schema: {type: string, const: 'private, no-store'}}}, content: {application/json: {schema: {$ref: '#/components/schemas/AiActivityList'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/ai/agent-runs:
    post:
      tags: [AI Intelligence]
      operationId: runSpecializedAiAgent
      description: Runs one allowlisted agent through the centralized gateway. The caller cannot select an arbitrary provider model schema tenant or actor. The returned result is a review-only suggestion and never mutates state.
      x-edt-agent-capabilities:
        connector.admin: [knowledge_ingestion, entity_resolution, technical_knowledge]
        scenario.create: [event_understanding]
        simulation.run: [causal_analysis, simulation_planning, prediction_explanation]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AiAgentRunRequest'}}}}
      responses:
        '202': {description: Real provider-backed structured result with provider usage audit and PENDING_REVIEW suggestion., headers: {Cache-Control: {schema: {type: string, const: 'private, no-store'}}}, content: {application/json: {schema: {$ref: '#/components/schemas/AiAgentRunResult'}}}}
        default: {$ref: '#/components/responses/Problem'}
...
  /v1/ai/suggestions:
    get:
      tags: [AI Intelligence]
      operationId: listAiSuggestions
      x-edt-authorization-any-of: [connector.admin, 'evidence.read.*']
      parameters:
        - {$ref: '#/components/parameters/AiPageSize'}
        - {name: review_decision, in: query, schema: {enum: [approve, reject]}}
      responses:
        '200': {description: Actor- and tenant-bound AI suggestions. A review record never changes suggestion status or mutates graph state., headers: {Cache-Control: {schema: {type: string, const: 'private, no-store'}}}, content: {application/json: {schema: {$ref: '#/components/schemas/AiSuggestionList'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/ai/suggestions/{suggestionId}/reviews:
    post:
      tags: [AI Intelligence]
      operationId: reviewAiSuggestion
      description: Records an authenticated human validation decision against the exact suggestion. An approval also persists an evidence-bound validated enterprise-memory record; it does not mutate the graph, simulation, or any external system.
      x-edt-required-capabilities: [connector.admin]
      parameters:
        - {name: suggestionId, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AiReviewRequest'}}}}
      responses:
        '201': {description: Immutable review receipt with mutation_performed false., headers: {Cache-Control: {schema: {type: string, const: 'private, no-store'}}}, content: {application/json: {schema: {$ref: '#/components/schemas/AiReviewReceipt'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/ai/learning/outcomes:
    post:
      tags: [AI Intelligence]
      operationId: recordValidatedAiLearningOutcome
      description: Persists an authenticated human-validated outcome for an approved suggestion as governed learning memory. The cited evidence is re-authorized server-side and the operation cannot mutate graph or simulation state.
      x-edt-required-capabilities: [connector.admin]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AiLearningOutcomeRequest'}}}}
      responses:
        '201': {description: Evidence-bound validated learning-memory receipt with both mutation flags false., headers: {Cache-Control: {schema: {type: string, const: 'private, no-store'}}}, content: {application/json: {schema: {$ref: '#/components/schemas/AiLearningOutcomeReceipt'}}}}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.enterprise-digital-twin.example/.well-known/openid-configuration
    contextHeader:
      type: apiKey
      in: header
      name: X-EDT-Context
      description: Opaque short-lived audience-bound server-issued membership context handle. Raw tenant identifiers are rejected.
    contextCookie:
      type: apiKey
      in: cookie
      name: EDT-Context
      description: Browser-only HTTP-only Secure SameSite opaque context handle; CSRF controls apply to commands.
  parameters:
...
    AiPageSize: {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
...
  responses:
    Problem:
      description: RFC 9457-compatible error with no unauthorized resource existence disclosure.
      content: {application/problem+json: {schema: {$ref: '../schemas/canonical.schema.json#/$defs/Problem'}}}
  schemas:
    AiStatus: {$ref: '../schemas/ai-intelligence.schema.json#/$defs/aiStatus'}
    AiActivityList: {$ref: '../schemas/ai-intelligence.schema.json#/$defs/activityList'}
    AiAgentRunRequest: {$ref: '../schemas/ai-intelligence.schema.json#/$defs/agentRunRequest'}
    AiAgentRunResult: {$ref: '../schemas/ai-intelligence.schema.json#/$defs/agentRunResult'}
...
    AiSuggestionList: {$ref: '../schemas/ai-intelligence.schema.json#/$defs/suggestionList'}
    AiReviewRequest: {$ref: '../schemas/ai-intelligence.schema.json#/$defs/reviewRequest'}
    AiReviewReceipt: {$ref: '../schemas/ai-intelligence.schema.json#/$defs/reviewReceipt'}
    AiLearningOutcomeRequest: {$ref: '../schemas/ai-intelligence.schema.json#/$defs/learningOutcomeRequest'}
    AiLearningOutcomeReceipt: {$ref: '../schemas/ai-intelligence.schema.json#/$defs/learningOutcomeReceipt'}
...
    SimulationRun: {$ref: '../schemas/canonical.schema.json#/$defs/SimulationRun'}
...
    SimulationRunAccepted:
      type: object
      additionalProperties: false
      required: [run, status_url, events_url]
      properties:
        run: {$ref: '#/components/schemas/SimulationRun'}
        status_url: {type: string, format: uri-reference}
        events_url: {type: string, format: uri-reference}
...